legacy/agents/agentB_tts/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
elevenlabs==0.2.26
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0

